refactor(akb-client): split 1070-line query-builder into focused modules
The query builder had grown to a single 1070-line file mixing the fluent
builder with a full PostgREST-style query compiler. Split it into six
single-responsibility modules with an acyclic dependency graph:

  query-builder.js (451)  builder classes + state/node construction (public)
  query-serialize.js (329) URL + AST serialization + boolean-expr parser
  query-compiler.js (146) GET/POST + rows/AST fallback decision + URL-safety
  query-result.js (67)    single()/maybeSingle() envelope unwrapping
  query-values.js (52)    scalar/filter value formatting (leaf)
  query-types.js (40)     shared JSDoc typedefs (no runtime code)

Also:
- Extract the 12 named-operator shortcut methods (eq/neq/.../not),
  previously duplicated verbatim across QueryBuilder and GroupBuilder,
  into a shared FilterBuilder base class.
- Remove the unused NAMED_OPERATORS constant (dead code).

Behavior is unchanged: the public surface is still the single
createQueryBuilder export, and the black-box builder tests pass without
modification (26/26), tsc --noEmit is clean, and codegen checks pass.
